moved shared model into a shared file
