+++ shared-data/python/opentrons_shared_data/protocol/models/__init__.py
@@ -2,7 +2,7 @@
 from . import protocol_schema_v6, protocol_schema_v7
 from .protocol_schema_v6 import ProtocolSchemaV6
 from .protocol_schema_v7 import ProtocolSchemaV7
-from .shared_models import Liquid, Labware
+from .shared_models import Liquid, Labware, CommandAnnotation, Location, ProfileStep, WellLocation, OffsetVector, Dimensions, GroupMetadata, Shape, WellDefinition, Metadata, Module, Pipette, Robot, DesignerApplication, Dimensions
 
 __all__ = [
     "ProtocolSchemaV6",
@@ -10,5 +10,7 @@ __all__ = [
     "ProtocolSchemaV7",
     "protocol_schema_v7",
     "Liquid",
-    "Labware"
+    "Labware",
+    "CommandAnnotation", "Location", "ProfileStep", "WellLocation", "OffsetVector",
+    "Dimensions", "GroupMetadata", "Shape", "WellDefinition", "Metadata", "Module", "Pipette", "Robot", "DesignerApplication"
 ]
